requirements.txt
Flask==2.0.1
Jinja2>=3.0.0
requests==2.26.0
beautifulsoup4==4.9.3
WeasyPrint>=52.5
pathlib>=1.0.1
logging>=0.5.1.2
python-tgpt[all]>=0.5.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=10.0.0
pytgpt>=1.0.0